Fix issues with copy of some files

+++ rapthor/lib/cwl.py
@@ -117,30 +117,52 @@ def is_cwl_file_or_directory(cwl_obj):
     return is_cwl_file(cwl_obj) or is_cwl_directory(cwl_obj)
 
 
-def copy_cwl_object(src_obj, dest_dir):
+def copy_cwl_object(src_obj, dest_dir, overwrite):
     """
     Copy a CWL file or directory object to the specified destination directory.
+
+    Parameters
+    ----------
+    src_obj : object
+        Source object of the copy
+    dest_dir: str
+        Path of destination directory to which src_obj will be copied
+    overwrite : bool
+        If True, existing files will be overwritten by a new copy. If False, existing
+        files are not overwritten
     """
     if is_cwl_file_or_directory(src_obj):
         os.makedirs(dest_dir, exist_ok=True)
         src = Path(src_obj['path'])
         dest = Path(dest_dir) / src.name
-        if is_cwl_file(src_obj):
-            shutil.copy(src, dest)
-        elif is_cwl_directory(src_obj):
-            shutil.copytree(src, dest, dirs_exist_ok=True)
+        if not os.path.exists(dest) or overwrite:
+            if is_cwl_file(src_obj):
+                shutil.copy(src, dest)
+            elif is_cwl_directory(src_obj):
+                shutil.copytree(src, dest, dirs_exist_ok=True)
     # Otherwise, do nothing
 
 
-def copy_cwl_recursive(src_obj, dest_dir):
+def copy_cwl_recursive(src_obj, dest_dir, overwrite):
     """
-    Recursively copy CWL file or directory objects to the specified destination directory.
+    Recursively copy CWL file or directory objects to the specified destination
+    directory.
+
+    Parameters
+    ----------
+    src_obj : object or list of objects
+        Source object(s) of the copy
+    dest_dir: str
+        Path of destination directory to which src_obj will be copied
+    overwrite : bool
+        If True, existing files will be overwritten by a new copy. If False, existing
+        files are not overwritten
     """
     if isinstance(src_obj, list):
         for item in src_obj:
-            copy_cwl_recursive(item, dest_dir)
+            copy_cwl_recursive(item, dest_dir, overwrite)
     elif is_cwl_file_or_directory(src_obj):
-        copy_cwl_object(src_obj, dest_dir)
+        copy_cwl_object(src_obj, dest_dir, overwrite)
     # Otherwise, do nothing
 
 
@@ -148,6 +170,11 @@ def remove_or_log_error(path: Path):
     """
     Remove a file or directory at the specified path.
     Log a warning if the file or directory does not exist.
+
+    Parameters
+    ----------
+    path: Path object
+        Path of file or directory to remove
     """
     try:
         if path.is_file():
@@ -157,9 +184,15 @@ def remove_or_log_error(path: Path):
     except FileNotFoundError:
         logger.warning("Cannot remove non-existing path: %s", path)
 
+
 def clean_if_cwl_file_or_directory(src_obj):
     """
     Remove CWL file or directory objects from the filesystem.
+
+    Parameters
+    ----------
+    src_obj : object or list of objects
+        Source object(s) to be removed
     """
     if isinstance(src_obj, list):
         for item in src_obj:
